utils: move lazy_abort_source from cloud_storage
I intend on moving some remote code out of cloud_storage into a new
subdirectory that doesn't depend on cloud_storage abstractions. To that
end, this pulls out cloud_storage::lazy_abort_source into the utilities
folder.
